atualização estrutural

+++ src/main/java/hotel/alura/rest/controller/ReservaController.java
@@ -82,7 +82,7 @@ public class ReservaController {
     }
     
     @PostMapping("/simular-valor")
-    public Long simularEstadia(@RequestBody ValorReservaDTO valores){
+    public Double simularEstadia(@RequestBody ValorReservaDTO valores){
         return service.valorEstadia(valores);
     } 
 }
